helpers/requirements.txt
pyserial
click
black